% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{cleans the LOCATION_NAME column}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{the name of the file to read (character)}
}
\value{
a data frame with a clean LOCATION_NAME column
}
\description{
cleans the LOCATION_NAME column by stripping out the country name
(including the colon) and converts names to title case (as opposed
to all caps).
}
\examples{
\dontrun{eq_clean <- eq_location_clean(eq_clean <- raw_data)}

}
