% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(eq_data, annot_col)
}
\arguments{
\item{eq_data}{a dataframe containing info on earthquakes. The data frame should have the following columns: LONGITUDE, LATITUDE, EQ_PRIMARY}

\item{annot_col}{the name of the column containing information to show on the map.}
}
\description{
maps earthquakes on a leaflet object
}
\examples{
\dontrun{readr::read_delim(eq_raw_file, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")}

}
