% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{data}
\name{theme_timeline}
\alias{theme_timeline}
\title{NOAA timeline theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 59.}
\usage{
theme_timeline
}
\description{
A simple theme fot the NOAACourser package
}
\examples{
\dontrun{ggplot() +
geom_timeline(data = eq_20e, aes(x = DATE, size = EQ_PRIMARY, y = COUNTRY, color = DEATHS)) +
scale_size_continuous(name  ="Richter scale value") +
scale_colour_continuous(name  ="# deaths") +
theme_timeline}

}
\keyword{datasets}
